requirements.txt
streamlit
tensorflow
numpy
Pillow
gtts
reportlab
